+++ test/test_kfac.py
@@ -3,7 +3,6 @@
 from typing import Dict, Iterable, List, Tuple, Union
 
 from einops.layers.torch import Rearrange
-from numpy import eye, random
 from pytest import mark
 from torch import (
     Tensor,
@@ -27,8 +26,8 @@ from torch.nn import (
 )
 
 from curvlinops import EFLinearOperator, GGNLinearOperator
-from curvlinops.examples.utils import report_nonclose
 from curvlinops.kfac import FisherType, KFACLinearOperator, KFACType
+from curvlinops.utils import allclose_report
 from test.cases import DEVICES, DEVICES_IDS
 from test.utils import (
     Conv2dModel,
@@ -43,6 +42,7 @@ from test.utils import (
     classification_targets,
     compare_consecutive_matmats,
     compare_matmat,
+    eye_like,
     maybe_exclude_or_shuffle_parameters,
     regression_targets,
 )
@@ -86,7 +86,7 @@ def test_kfac_type2(
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
     )
-    kfac_torch = KFACLinearOperator(
+    kfac = KFACLinearOperator(
         model,
         loss_func,
         params,
@@ -95,14 +95,13 @@ def test_kfac_type2(
         fisher_type=FisherType.TYPE2,
         separate_weight_and_bias=separate_weight_and_bias,
     )
-    kfac = kfac_torch.to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    kfac_mat = kfac @ eye_like(kfac)
 
-    report_nonclose(ggn, kfac_mat)
+    assert allclose_report(ggn, kfac_mat)
 
     # Check that input covariances were not computed
     if exclude == "weight":
-        assert len(kfac_torch._input_covariances) == 0
+        assert len(kfac._input_covariances) == 0
 
 
 @mark.parametrize("setting", [KFACType.EXPAND, KFACType.REDUCE])
@@ -155,7 +154,7 @@ def test_kfac_type2_weight_sharing(
         batch_size_fn=batch_size_fn,
         separate_weight_and_bias=separate_weight_and_bias,
     )
-    kfac_torch = KFACLinearOperator(
+    kfac = KFACLinearOperator(
         model,
         loss_func,
         params,
@@ -165,14 +164,13 @@ def test_kfac_type2_weight_sharing(
         kfac_approx=setting,  # choose KFAC approximation consistent with setting
         separate_weight_and_bias=separate_weight_and_bias,
     )
-    kfac = kfac_torch.to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    kfac_mat = kfac @ eye_like(kfac)
 
-    report_nonclose(ggn, kfac_mat, rtol=1e-4)
+    assert allclose_report(ggn, kfac_mat, rtol=1e-4)
 
     # Check that input covariances were not computed
     if exclude == "weight":
-        assert len(kfac_torch._input_covariances) == 0
+        assert len(kfac._input_covariances) == 0
 
 
 @mark.parametrize(
@@ -222,13 +220,13 @@ def test_kfac_mc(
         fisher_type=FisherType.MC,
         mc_samples=2_000,
         separate_weight_and_bias=separate_weight_and_bias,
-    ).to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    )
+    kfac_mat = kfac @ eye_like(kfac)
 
     atol = {"sum": 5e-1, "mean": 1e-2}[loss_func.reduction]
     rtol = {"sum": 2e-2, "mean": 2e-2}[loss_func.reduction]
 
-    report_nonclose(ggn, kfac_mat, rtol=rtol, atol=atol)
+    assert allclose_report(ggn, kfac_mat, rtol=rtol, atol=atol)
 
 
 @mark.parametrize(
@@ -287,13 +285,13 @@ def test_kfac_mc_weight_sharing(
         mc_samples=2_000,
         kfac_approx=setting,  # choose KFAC approximation consistent with setting
         separate_weight_and_bias=separate_weight_and_bias,
-    ).to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    )
+    kfac_mat = kfac @ eye_like(kfac)
 
     atol = {"sum": 5e-1, "mean": 1e-2}[loss_func.reduction]
     rtol = {"sum": 2e-2, "mean": 2e-2}[loss_func.reduction]
 
-    report_nonclose(ggn, kfac_mat, rtol=rtol, atol=atol)
+    assert allclose_report(ggn, kfac_mat, rtol=rtol, atol=atol)
 
 
 @mark.parametrize(
@@ -334,10 +332,10 @@ def test_kfac_one_datum(
         batch_size_fn=batch_size_fn,
         fisher_type=FisherType.TYPE2,
         separate_weight_and_bias=separate_weight_and_bias,
-    ).to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    )
+    kfac_mat = kfac @ eye_like(kfac)
 
-    report_nonclose(ggn, kfac_mat)
+    assert allclose_report(ggn, kfac_mat)
 
 
 @mark.parametrize(
@@ -379,13 +377,13 @@ def test_kfac_mc_one_datum(
         fisher_type=FisherType.MC,
         mc_samples=11_000,
         separate_weight_and_bias=separate_weight_and_bias,
-    ).to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    )
+    kfac_mat = kfac @ eye_like(kfac)
 
     atol = {"sum": 1e-3, "mean": 1e-3}[loss_func.reduction]
     rtol = {"sum": 3e-2, "mean": 3e-2}[loss_func.reduction]
 
-    report_nonclose(ggn, kfac_mat, rtol=rtol, atol=atol)
+    assert allclose_report(ggn, kfac_mat, rtol=rtol, atol=atol)
 
 
 @mark.parametrize(
@@ -427,10 +425,10 @@ def test_kfac_ef_one_datum(
         batch_size_fn=batch_size_fn,
         fisher_type=FisherType.EMPIRICAL,
         separate_weight_and_bias=separate_weight_and_bias,
-    ).to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    )
+    kfac_mat = kfac @ eye_like(kfac)
 
-    report_nonclose(ef, kfac_mat)
+    assert allclose_report(ef, kfac_mat)
 
 
 @mark.parametrize("dev", DEVICES, ids=DEVICES_IDS)
@@ -500,14 +498,8 @@ def test_multi_dim_output(
 
     # KFAC for deep linear network with 4d input and output
     params = list(model.parameters())
-    kfac = KFACLinearOperator(
-        model,
-        loss_func,
-        params,
-        data,
-        fisher_type=fisher_type,
-    ).to_scipy()
-    kfac_mat = kfac @ eye(kfac.shape[1])
+    kfac = KFACLinearOperator(model, loss_func, params, data, fisher_type=fisher_type)
+    kfac_mat = kfac @ eye_like(kfac)
 
     # KFAC for deep linear network with 4d input and equivalent 2d output
     manual_seed(711)
@@ -531,10 +523,10 @@ def test_multi_dim_output(
         params_flat,
         data_flat,
         fisher_type=fisher_type,
-    ).to_scipy()
-    kfac_flat_mat = kfac_flat @ eye(kfac_flat.shape[1])
+    )
+    kfac_flat_mat = kfac_flat @ eye_like(kfac_flat)
 
-    report_nonclose(kfac_mat, kfac_flat_mat)
+    assert allclose_report(kfac_mat, kfac_flat_mat)
 
 
 @mark.parametrize("fisher_type", KFACLinearOperator._SUPPORTED_FISHER_TYPE)
@@ -581,14 +573,13 @@ def test_expand_setting_scaling(
 
     # KFAC with sum reduction
     loss_func = loss(reduction="sum").to(dev)
-    kfac_sum_torch = KFACLinearOperator(
+    kfac_sum = KFACLinearOperator(
         model,
         loss_func,
         params,
         data,
         fisher_type=fisher_type,
     )
-    kfac_sum = kfac_sum_torch.to_scipy()
     # FOOF does not scale the gradient covariances, even when using a mean reduction
     if fisher_type != FisherType.FORWARD_ONLY:
         # Simulate a mean reduction by manually scaling the gradient covariances
@@ -597,9 +588,9 @@ def test_expand_setting_scaling(
             output_random_variable_size = 3
             # MSE loss averages over number of output channels
             loss_term_factor *= output_random_variable_size
-        for ggT in kfac_sum_torch._gradient_covariances.values():
-            ggT.div_(kfac_sum_torch._N_data * loss_term_factor)
-    kfac_simulated_mean_mat = kfac_sum @ eye(kfac_sum.shape[1])
+        for ggT in kfac_sum._gradient_covariances.values():
+            ggT.div_(kfac_sum._N_data * loss_term_factor)
+    kfac_simulated_mean_mat = kfac_sum @ eye_like(kfac_sum)
 
     # KFAC with mean reduction
     loss_func = loss(reduction="mean").to(dev)
@@ -609,10 +600,10 @@ def test_expand_setting_scaling(
         params,
         data,
         fisher_type=fisher_type,
-    ).to_scipy()
-    kfac_mean_mat = kfac_mean @ eye(kfac_mean.shape[1])
+    )
+    kfac_mean_mat = kfac_mean @ eye_like(kfac_mean)
 
-    report_nonclose(kfac_simulated_mean_mat, kfac_mean_mat)
+    assert allclose_report(kfac_simulated_mean_mat, kfac_mean_mat)
 
 
 @mark.parametrize("separate_weight_and_bias", [False], ids=["joint_bias"])
@@ -823,7 +814,7 @@ def test_forward_only_fisher_type(
         foof_simulated._gradient_covariances[name] = torch_eye(
             block.shape[0], dtype=block.dtype, device=block.device
         )
-    simulated_foof_mat = foof_simulated.to_scipy() @ eye(foof_simulated.shape[1])
+    simulated_foof_mat = foof_simulated @ eye_like(foof_simulated)
 
     # Compute KFAC with `fisher_type=FisherType.FORWARD_ONLY`
     foof = KFACLinearOperator(
@@ -835,12 +826,12 @@ def test_forward_only_fisher_type(
         separate_weight_and_bias=separate_weight_and_bias,
         fisher_type=FisherType.FORWARD_ONLY,
     )
-    foof_mat = foof.to_scipy() @ eye(foof.shape[1])
+    foof_mat = foof @ eye_like(foof)
 
     # Check for equivalence
     assert len(foof_simulated._input_covariances) == len(foof._input_covariances)
     assert len(foof_simulated._gradient_covariances) == len(foof._gradient_covariances)
-    report_nonclose(simulated_foof_mat, foof_mat)
+    assert allclose_report(simulated_foof_mat, foof_mat)
 
     # Check that input covariances were not computed
     if exclude == "weight":
@@ -905,7 +896,7 @@ def test_forward_only_fisher_type_exact_case(
         separate_weight_and_bias=separate_weight_and_bias,
         fisher_type=FisherType.FORWARD_ONLY,
     )
-    foof_mat = foof.to_scipy() @ eye(foof.shape[1])
+    foof_mat = foof @ eye_like(foof)
 
     # Check for equivalence
     num_data = sum(X.shape[0] for X, _ in data)
@@ -913,7 +904,7 @@ def test_forward_only_fisher_type_exact_case(
     out_dim = y.shape[1]
     # See the docstring for the explanation of the scale
     scale = num_data if loss_func.reduction == "sum" else 1 / out_dim
-    report_nonclose(ggn, 2 * scale * foof_mat)
+    assert allclose_report(ggn, 2 * scale * foof_mat)
 
     # Check that input covariances were not computed
     if exclude == "weight":
@@ -1001,7 +992,7 @@ def test_forward_only_fisher_type_exact_weight_sharing_case(
         kfac_approx=setting,  # choose KFAC approximation consistent with setting
         separate_weight_and_bias=separate_weight_and_bias,
     )
-    foof_mat = foof.to_scipy() @ eye(foof.shape[1])
+    foof_mat = foof @ eye_like(foof)
 
     # Check for equivalence
     num_data = sum(X.shape[0] for X, _ in data)
@@ -1016,7 +1007,7 @@ def test_forward_only_fisher_type_exact_weight_sharing_case(
             else X.shape[1:-1].numel()
         )
         scale *= sequence_length
-    report_nonclose(ggn, 2 * scale * foof_mat, rtol=1e-4)
+    assert allclose_report(ggn, 2 * scale * foof_mat, rtol=1e-4)
 
     # Check that input covariances were not computed
     if exclude == "weight":
@@ -1093,7 +1084,6 @@ def test_bug_132_dtype_deterministic_checks(dev: device):
     """
     # make deterministic
     manual_seed(0)
-    random.seed(0)
 
     # create a toy problem, load everything to float64
     dt = float64